refacto, add security
