+++ templates/security/account.html.twig
@@ -1,4 +1,4 @@
- {% extends 'base.html.twig' %}  
+{% extends 'base.html.twig' %}
 
 {% block title %}Register{% endblock %}
 
@@ -16,7 +16,7 @@
             {{ form_row(registrationForm.agreeTerms) }}
 
             <div class="row">
-                <button class="btn btn-primary mx-auto">Register</button>
+                <button class="btn btn-primary mx-auto">Save</button>
             </div>
             {{ form_end(registrationForm) }}
         </div>
