--- templates/security/admin.html.twig
@@ -1,1 +1,0 @@
-test
