+++ templates/videogame/showVideogame.html.twig
@@ -1,5 +1,5 @@
 {% extends 'base.html.twig' %}
-{% block title %}Save Videogame{% endblock %}
+{% block title %}{{ title }}{% endblock %}
 
 {% block body %}
